+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to the "Project Timer" extension will be documented in this 
 ### Fixed
 - Fixed unexpected horizontal scrollbar in statistics page chart.
 
+### Improved
+- Added icon to statistics page panel.
+
 ## [0.5.3] - 2026-05-05
 ### Fixed
 - Fixed build configuration which wrongly package asset files.